Refactors database documents into seperate files
