+++ ios/Flutter/flutter_export_environment.sh
@@ -5,6 +5,6 @@ export "FLUTTER_APPLICATION_PATH=/Users/samjakob/Desktop/ApolloTV/kamino"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
-export "FLUTTER_FRAMEWORK_DIR=/Users/samjakob/development/flutter/bin/cache/artifacts/engine/ios"
+export "FLUTTER_FRAMEWORK_DIR=/Users/samjakob/development/flutter/bin/cache/artifacts/engine/ios-release"
 export "FLUTTER_BUILD_NAME=1.4.8"
-export "FLUTTER_BUILD_NUMBER=104080"
+export "FLUTTER_BUILD_NUMBER=104083"
