+++ lib/interface/content/movie_layout.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:kamino/generated/i18n.dart';
-import 'package:kamino/main.dart';
 import 'package:kamino/models/content/movie.dart';
 
 class MovieLayout {
