+++ lib/partials/content_poster.dart
@@ -2,7 +2,6 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:kamino/external/api/tmdb.dart';
 import 'package:kamino/models/content/content.dart';
-import 'package:kamino/res/bottom_gradient.dart';
 import 'package:shimmer/shimmer.dart';
 
 class ContentPoster extends StatefulWidget {
